fix up some type inconsistencies
fix more type inconsistencies, allow mixed types between float + int in from_security_mappings

fix more type inconsistencies

+++ src/backtest_lib/engine/execute/perfect_world.py
@@ -46,7 +46,7 @@ logger = logging.getLogger(__name__)
 
 @dataclass(slots=True)
 class _ValueReallocation:
-    inner: UniverseMapping[int]
+    inner: UniverseMapping[float]
 
 
 @dataclass(slots=True)
